requirements.txt
Django==3.2.5
asgiref==3.4.1
dj-database-url==0.5.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
django-csp==3.7
django-permissions-policy==4.1.0
django-heroku==0.3.1
